Works very well. (nonconvex star obstacle)
./gf 512 512 1000 0.0001 4000 0.004 0.03 4 0.1 1

Notes: the tolerance has to be set very low to have a mass stuck and the star. I think this is a correct solution. The problem is if the tolerance is high then the mass that is supposed to be stuck disappears over time.

Using small lambda which results in better performance.